Refactors and minor fixes of tests
