+++ automation/src/test/java/com/microsoft/identity/common/test/automation/AcquireTokenATExpiryTest.java
@@ -1,33 +1,15 @@
-//  Copyright (c) Microsoft Corporation.
-//  All rights reserved.
-//
-//  This code is licensed under the MIT License.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files(the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions :
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
 package com.microsoft.identity.common.test.automation;
 
 import com.microsoft.identity.common.test.automation.actors.User;
 import com.microsoft.identity.common.test.automation.interactions.ClickDone;
+import com.microsoft.identity.common.test.automation.model.TokenCacheItemReadResult;
+import com.microsoft.identity.common.test.automation.questions.AccessToken;
 import com.microsoft.identity.common.test.automation.questions.ExpectedCacheItemCount;
 import com.microsoft.identity.common.test.automation.questions.TokenCacheItemCount;
+import com.microsoft.identity.common.test.automation.questions.TokenCacheItemFromResult;
 import com.microsoft.identity.common.test.automation.tasks.AcquireToken;
+import com.microsoft.identity.common.test.automation.tasks.AcquireTokenSilent;
+import com.microsoft.identity.common.test.automation.tasks.ExpireAccessToken;
 import com.microsoft.identity.common.test.automation.tasks.ReadCache;
 import com.microsoft.identity.common.test.automation.utility.Scenario;
 import com.microsoft.identity.common.test.automation.utility.TestConfigurationQuery;
@@ -36,6 +18,7 @@ import net.serenitybdd.junit.runners.SerenityParameterizedRunner;
 import net.serenitybdd.screenplay.abilities.BrowseTheWeb;
 import net.thucydides.core.annotations.Managed;
 import net.thucydides.core.annotations.Steps;
+import net.thucydides.core.annotations.WithTag;
 import net.thucydides.junit.annotations.TestData;
 
 import org.junit.AfterClass;
@@ -51,15 +34,18 @@ import java.util.Collection;
 
 import io.appium.java_client.service.local.AppiumDriverLocalService;
 
+import static net.serenitybdd.screenplay.GivenWhenThen.givenThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.seeThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.then;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.not;
 
 @RunWith(SerenityParameterizedRunner.class)
-public class AcquireTokenBasicPromptAlwaysTest {
+@WithTag("requires:none")
+public class AcquireTokenATExpiryTest {
 
     @TestData
-    public static Collection<Object[]> FederationProviders() {
+    public static Collection<Object[]> FederationProviders(){
 
 
         return Arrays.asList(new Object[][]{
@@ -73,9 +59,24 @@ public class AcquireTokenBasicPromptAlwaysTest {
 
     }
 
+    @Steps
+    AcquireTokenSilent acquireTokenSilent;
+
+    @Steps
+    AcquireToken acquireToken;
+
+    @Steps
+    ExpireAccessToken expireAccessToken;
+
+    @Steps
+    ReadCache readCache;
+
+    @Steps
+    ClickDone clickDone;
+
     static AppiumDriverLocalService appiumService = null;
 
-    @Managed(driver = "Appium")
+    @Managed(driver="Appium")
     WebDriver hisMobileDevice;
 
     @BeforeClass
@@ -92,21 +93,12 @@ public class AcquireTokenBasicPromptAlwaysTest {
     private User james;
     private String federationProvider;
 
-    @Steps
-    AcquireToken acquireToken;
-
-    @Steps
-    ReadCache readCache;
-
-    @Steps
-    ClickDone clickDone;
-
-    public AcquireTokenBasicPromptAlwaysTest(String federationProvider) {
+    public AcquireTokenATExpiryTest(String federationProvider){
         this.federationProvider = federationProvider;
     }
 
     @Before
-    public void jamesCanUseAMobileDevice() {
+    public void jamesCanUseAMobileDevice(){
         TestConfigurationQuery query = new TestConfigurationQuery();
         query.federationProvider = this.federationProvider;
         query.isFederated = true;
@@ -115,13 +107,14 @@ public class AcquireTokenBasicPromptAlwaysTest {
         james.can(BrowseTheWeb.with(hisMobileDevice));
     }
 
-    private static User getUser(TestConfigurationQuery query) {
+    private static User getUser(TestConfigurationQuery query){
 
         Scenario scenario = Scenario.GetScenario(query);
 
         User newUser = User.named("james");
         newUser.setFederationProvider(scenario.getTestConfiguration().getUsers().getFederationProvider());
         newUser.setTokenRequest(scenario.getTokenRequest());
+        newUser.setSilentTokenRequest(scenario.getSilentTokenRequest());
         newUser.setCredential(scenario.getCredential());
 
         return newUser;
@@ -129,15 +122,30 @@ public class AcquireTokenBasicPromptAlwaysTest {
 
 
     @Test
-    public void should_be_able_to_acquire_token() {
+    public void should_be_able_to_new_access_token_after_expiry_on_silent() {
 
-        james.attemptsTo(
-                acquireToken.withPrompt("Always"),
+        givenThat(james).wasAbleTo(
+                acquireToken,
+                clickDone,
+                readCache);
+        int expectedCacheCount = james.asksFor(ExpectedCacheItemCount.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
+
+        String accessToken1 = james.asksFor(AccessToken.displayed());
+        TokenCacheItemReadResult cacheItem = james.asksFor(TokenCacheItemFromResult.displayed());
+
+        givenThat(james).wasAbleTo(
+                clickDone,
+                expireAccessToken.withTokenCacheItem(cacheItem),
+                clickDone,
+                acquireToken
+                        .tokenExists(true)
+                        .withUserIdentifier(cacheItem.displayableId),
                 clickDone,
                 readCache);
 
-        int expectedCacheCount = james.asksFor(ExpectedCacheItemCount.displayed());
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
+        then(james).should(seeThat(AccessToken.displayed(), not(accessToken1)));
+
 
     }
