+++ automation/src/test/java/com/microsoft/identity/common/test/automation/AcquireTokenSilentATAndRTExpiryTest.java
@@ -1,43 +1,26 @@
-//  Copyright (c) Microsoft Corporation.
-//  All rights reserved.
-//
-//  This code is licensed under the MIT License.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files(the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions :
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
 package com.microsoft.identity.common.test.automation;
 
 import com.microsoft.identity.common.test.automation.actors.User;
 import com.microsoft.identity.common.test.automation.interactions.ClickDone;
 import com.microsoft.identity.common.test.automation.model.Constants;
+import com.microsoft.identity.common.test.automation.model.TokenCacheItemReadResult;
+import com.microsoft.identity.common.test.automation.questions.ADALError;
 import com.microsoft.identity.common.test.automation.questions.ExpectedCacheItemCount;
 import com.microsoft.identity.common.test.automation.questions.ExpectedCacheItemCountWithFoci;
 import com.microsoft.identity.common.test.automation.questions.TokenCacheItemCount;
+import com.microsoft.identity.common.test.automation.questions.TokenCacheItemFromResult;
 import com.microsoft.identity.common.test.automation.tasks.AcquireToken;
 import com.microsoft.identity.common.test.automation.tasks.AcquireTokenSilent;
 import com.microsoft.identity.common.test.automation.tasks.ClearCache;
+import com.microsoft.identity.common.test.automation.tasks.ExpireATAndInvalidateRT;
 import com.microsoft.identity.common.test.automation.tasks.ReadCache;
+import com.microsoft.identity.common.test.automation.ui.Results;
 import com.microsoft.identity.common.test.automation.utility.Scenario;
 import com.microsoft.identity.common.test.automation.utility.TestConfigurationQuery;
 
 import net.serenitybdd.junit.runners.SerenityParameterizedRunner;
 import net.serenitybdd.screenplay.abilities.BrowseTheWeb;
+import net.serenitybdd.screenplay.waits.WaitUntil;
 import net.thucydides.core.annotations.Managed;
 import net.thucydides.core.annotations.Steps;
 import net.thucydides.core.annotations.WithTag;
@@ -56,18 +39,20 @@ import java.util.Collection;
 
 import io.appium.java_client.service.local.AppiumDriverLocalService;
 
+import static net.serenitybdd.screenplay.GivenWhenThen.andThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.givenThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.seeThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.then;
 import static net.serenitybdd.screenplay.GivenWhenThen.when;
+import static net.serenitybdd.screenplay.matchers.WebElementStateMatchers.isVisible;
 import static org.hamcrest.Matchers.is;
 
 @RunWith(SerenityParameterizedRunner.class)
 @WithTag("requires:none")
-public class AcquireTokenInteractiveThenSilentBasicTest {
+public class AcquireTokenSilentATAndRTExpiryTest {
 
     @TestData
-    public static Collection<Object[]> FederationProviders() {
+    public static Collection<Object[]> FederationProviders(){
 
 
         return Arrays.asList(new Object[][]{
@@ -88,19 +73,20 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
     AcquireToken acquireToken;
 
     @Steps
-    ReadCache readCache;
+    ExpireATAndInvalidateRT expireATAndInvalidateRT;
 
     @Steps
-    ClearCache clearCache;
+    ReadCache readCache;
 
     @Steps
     ClickDone clickDone;
 
+    @Steps
+    ClearCache clearCache;
+
     static AppiumDriverLocalService appiumService = null;
 
-    private String mSecondaryResource;
-
-    @Managed(driver = "Appium")
+    @Managed(driver="Appium")
     WebDriver hisMobileDevice;
 
     @BeforeClass
@@ -117,12 +103,12 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
     private User james;
     private String federationProvider;
 
-    public AcquireTokenInteractiveThenSilentBasicTest(String federationProvider) {
+    public AcquireTokenSilentATAndRTExpiryTest(String federationProvider){
         this.federationProvider = federationProvider;
     }
 
     @Before
-    public void jamesCanUseAMobileDevice() {
+    public void jamesCanUseAMobileDevice(){
         TestConfigurationQuery query = new TestConfigurationQuery();
         query.federationProvider = this.federationProvider;
         query.isFederated = true;
@@ -131,10 +117,10 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
         james.can(BrowseTheWeb.with(hisMobileDevice));
     }
 
-    private User getUser(TestConfigurationQuery query) {
+    private static User getUser(TestConfigurationQuery query){
 
         Scenario scenario = Scenario.GetScenario(query);
-        mSecondaryResource = scenario.getTestConfiguration().getResourceIds().get(1);
+
         User newUser = User.named("james");
         newUser.setFederationProvider(scenario.getTestConfiguration().getUsers().getFederationProvider());
         newUser.setTokenRequest(scenario.getTokenRequest());
@@ -146,7 +132,7 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
 
 
     @Test
-    public void should_be_able_to_acquire_token_and_then_acquire_silent() {
+    public void should_not_be_able_access_token_after_at_rt_expiry_on_silent_unique_id() {
 
         //clear cache
         givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
@@ -157,48 +143,58 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
                 acquireToken,
                 clickDone,
                 readCache);
-        int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCount.displayed());
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+        int expectedCacheCount = james.asksFor(ExpectedCacheItemCount.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
 
-        when(james).attemptsTo(
-                clickDone,
-                acquireTokenSilent,
-                clickDone,
-                readCache
-        );
+        TokenCacheItemReadResult cacheItem = james.asksFor(TokenCacheItemFromResult.displayed());
 
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+        givenThat(james).wasAbleTo(
+                clickDone,
+                expireATAndInvalidateRT.withTokenCacheItem(cacheItem),
+                clickDone,
+                acquireTokenSilent.withUniqueId(cacheItem.uniqueUserId),
+                WaitUntil.the(Results.RESULT_FIELD, isVisible()).forNoMoreThan(10).seconds());
+
+        then(james).should(seeThat(ADALError.displayed(), is(com.microsoft.identity.common.adal.error.ADALError.AUTH_REFRESH_FAILED_PROMPT_NOT_ALLOWED.name())));
+
+        when(james).attemptsTo(clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
 
     }
 
     @Test
-    public void should_be_able_to_acquire_token_and_then_acquire_silent_with_different_resource() {
+    public void should_not_be_able_access_token_after_at_rt_expiry_on_silent_displayable_id() {
 
         //clear cache
         givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
         then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
+
 
         givenThat(james).wasAbleTo(
                 clickDone,
                 acquireToken,
                 clickDone,
                 readCache);
-        int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCount.displayed());
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+        int expectedCacheCount = james.asksFor(ExpectedCacheItemCount.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
 
-        when(james).attemptsTo(
+        TokenCacheItemReadResult cacheItem = james.asksFor(TokenCacheItemFromResult.displayed());
+        givenThat(james).wasAbleTo(
                 clickDone,
-                acquireTokenSilent.withResourceId(mSecondaryResource),
+                expireATAndInvalidateRT.withTokenCacheItem(cacheItem),
                 clickDone,
-                readCache
-        );
-        expectedCacheCountToken  += expectedCacheCountToken/2;
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+                acquireTokenSilent.withUniqueId(cacheItem.displayableId),
+                WaitUntil.the(Results.RESULT_FIELD, isVisible()).forNoMoreThan(10).seconds());
+
+        then(james).should(seeThat(ADALError.displayed(), is(com.microsoft.identity.common.adal.error.ADALError.AUTH_REFRESH_FAILED_PROMPT_NOT_ALLOWED.name())));
+
+        when(james).attemptsTo(clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
 
     }
 
     @Test
-    public void should_be_able_to_acquire_token_and_then_acquire_silent_frt() {
+    public void should_not_be_able_access_token_after_at_frt_expiry_on_silent() {
 
         //clear cache
         givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
@@ -211,23 +207,46 @@ public class AcquireTokenInteractiveThenSilentBasicTest {
                         .withRedirectUri(Constants.OUTLOOK_REDIRECT_URI),
                 clickDone,
                 readCache);
+
         int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCountWithFoci.displayed());
         then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
 
-        when(james).attemptsTo(
+        givenThat(james).wasAbleTo(
                 clickDone,
                 acquireTokenSilent
                         .withClientId(Constants.ONE_DRIVE_CLIENT_ID)
                         .withRedirectUri(Constants.ONE_DRIVE_REDIRECT_URI),
                 clickDone,
-                readCache
-        );
+                readCache);
 
         int expectedFinalCacheCountCount = james.asksFor(ExpectedCacheItemCount.displayed()) + expectedCacheCountToken;
         then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedFinalCacheCountCount)));
 
-    }
+        TokenCacheItemReadResult cacheItem = james.asksFor(TokenCacheItemFromResult.displayed());
 
+        givenThat(james).wasAbleTo(
+                clickDone,
+                expireATAndInvalidateRT
+                        .withTokenCacheItem(cacheItem)
+                        .withClientId(Constants.OUTLOOK_CLIENT_ID),
+                clickDone);
+
+        andThat(james).wasAbleTo(
+                expireATAndInvalidateRT
+                        .withTokenCacheItem(cacheItem)
+                        .withClientId(Constants.ONE_DRIVE_CLIENT_ID),
+                clickDone);
+
+        when(james).attemptsTo(
+                acquireTokenSilent.withUniqueId(cacheItem.uniqueUserId),
+                WaitUntil.the(Results.RESULT_FIELD, isVisible()).forNoMoreThan(10).seconds());
+
+        then(james).should(seeThat(ADALError.displayed(), is(com.microsoft.identity.common.adal.error.ADALError.AUTH_REFRESH_FAILED_PROMPT_NOT_ALLOWED.name())));
+
+        when(james).attemptsTo(clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is((expectedCacheCountToken == 20) ? 12 : 8)));
+
+    }
 
 
 }
