+++ automation/src/test/java/com/microsoft/identity/common/test/automation/AcquireTokenSilentTest.java
@@ -25,10 +25,14 @@ package com.microsoft.identity.common.test.automation;
 
 import com.microsoft.identity.common.test.automation.actors.User;
 import com.microsoft.identity.common.test.automation.interactions.ClickDone;
+import com.microsoft.identity.common.test.automation.model.Constants;
 import com.microsoft.identity.common.test.automation.questions.AccessToken;
 import com.microsoft.identity.common.test.automation.questions.ExpectedCacheItemCount;
+import com.microsoft.identity.common.test.automation.questions.ExpectedCacheItemCountWithFoci;
 import com.microsoft.identity.common.test.automation.questions.TokenCacheItemCount;
 import com.microsoft.identity.common.test.automation.tasks.AcquireToken;
+import com.microsoft.identity.common.test.automation.tasks.AcquireTokenSilent;
+import com.microsoft.identity.common.test.automation.tasks.ClearCache;
 import com.microsoft.identity.common.test.automation.tasks.ReadCache;
 import com.microsoft.identity.common.test.automation.utility.Scenario;
 import com.microsoft.identity.common.test.automation.utility.TestConfigurationQuery;
@@ -37,6 +41,7 @@ import net.serenitybdd.junit.runners.SerenityParameterizedRunner;
 import net.serenitybdd.screenplay.abilities.BrowseTheWeb;
 import net.thucydides.core.annotations.Managed;
 import net.thucydides.core.annotations.Steps;
+import net.thucydides.core.annotations.WithTag;
 import net.thucydides.junit.annotations.TestData;
 
 import org.junit.AfterClass;
@@ -52,21 +57,19 @@ import java.util.Collection;
 
 import io.appium.java_client.service.local.AppiumDriverLocalService;
 
-import static net.serenitybdd.screenplay.GivenWhenThen.and;
 import static net.serenitybdd.screenplay.GivenWhenThen.givenThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.seeThat;
 import static net.serenitybdd.screenplay.GivenWhenThen.then;
 import static net.serenitybdd.screenplay.GivenWhenThen.when;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.not;
 
-/**
- * Test case : https://identitydivision.visualstudio.com/IDDP/DevEx-Client-SDK/_testManagement?planId=345909&suiteId=345910&_a=tests
- */
 @RunWith(SerenityParameterizedRunner.class)
-public class AcquireTokenPromptAutoThenAlwaysTest {
+@WithTag("requires:none")
+public class AcquireTokenSilentTest {
+
     @TestData
     public static Collection<Object[]> FederationProviders() {
+
 
         return Arrays.asList(new Object[][]{
                 {"ADFSv2"},
@@ -76,9 +79,27 @@ public class AcquireTokenPromptAutoThenAlwaysTest {
                 {"Shibboleth"}
 
         });
+
     }
 
+    @Steps
+    AcquireTokenSilent acquireTokenSilent;
+
+    @Steps
+    AcquireToken acquireToken;
+
+    @Steps
+    ReadCache readCache;
+
+    @Steps
+    ClearCache clearCache;
+
+    @Steps
+    ClickDone clickDone;
+
     static AppiumDriverLocalService appiumService = null;
+
+    private String mSecondaryResource;
 
     @Managed(driver = "Appium")
     WebDriver hisMobileDevice;
@@ -97,17 +118,7 @@ public class AcquireTokenPromptAutoThenAlwaysTest {
     private User james;
     private String federationProvider;
 
-    @Steps
-    AcquireToken acquireToken;
-
-    @Steps
-    ReadCache readCache;
-
-    @Steps
-    ClickDone clickDone;
-
-
-    public AcquireTokenPromptAutoThenAlwaysTest(String federationProvider) {
+    public AcquireTokenSilentTest(String federationProvider) {
         this.federationProvider = federationProvider;
     }
 
@@ -121,34 +132,140 @@ public class AcquireTokenPromptAutoThenAlwaysTest {
         james.can(BrowseTheWeb.with(hisMobileDevice));
     }
 
-    private static User getUser(TestConfigurationQuery query) {
+    private User getUser(TestConfigurationQuery query) {
+
         Scenario scenario = Scenario.GetScenario(query);
+        mSecondaryResource = scenario.getTestConfiguration().getResourceIds().get(1);
         User newUser = User.named("james");
         newUser.setFederationProvider(scenario.getTestConfiguration().getUsers().getFederationProvider());
         newUser.setTokenRequest(scenario.getTokenRequest());
+        newUser.setSilentTokenRequest(scenario.getSilentTokenRequest());
         newUser.setCredential(scenario.getCredential());
+
         return newUser;
     }
 
+
     @Test
-    public void should_be_able_to_acquire_token_refresh_auto() {
+    public void should_be_able_to_acquire_token_silent() {
+
+        james.attemptsTo(acquireTokenSilent);
+
+    }
+
+    @Test
+    public void should_be_able_to_acquire_token_and_then_acquire_silent() {
+
+        //clear cache
+        givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
+
         givenThat(james).wasAbleTo(
-                acquireToken.withPrompt("Auto").withUserIdentifier(james.getTokenRequest().getUserIdentitfier()),
+                clickDone,
+                acquireToken,
                 clickDone,
                 readCache);
+        int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCount.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+
+        when(james).attemptsTo(
+                clickDone,
+                acquireTokenSilent,
+                clickDone,
+                readCache
+        );
+
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+
+    }
+
+    @Test
+    public void should_be_able_to_acquire_token_and_then_acquire_silent_with_login_hint() {
+
+        //clear cache
+        givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
+
+        givenThat(james).wasAbleTo(
+                clickDone,
+                acquireToken,
+                clickDone,
+                readCache);
+
         int expectedCacheCount = james.asksFor(ExpectedCacheItemCount.displayed());
         then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
 
-        //Store the access token temporarily to compare later
         String accessToken1 = james.asksFor(AccessToken.displayed());
+
         james.attemptsTo(clickDone);
 
         when(james).attemptsTo(
-                acquireToken.withPrompt("Always").withUserIdentifier(null),
+                acquireTokenSilent.withUserIdentifier(james.getCredential().userName),
                 clickDone,
                 readCache);
-        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCount)));
-        and(james).should(seeThat(AccessToken.displayed(), not(accessToken1)));
+
+        then(james).should(seeThat(AccessToken.displayed(), is(accessToken1)));
+
 
     }
+
+    @Test
+    public void should_be_able_to_acquire_token_and_then_acquire_silent_with_different_resource() {
+
+        //clear cache
+        givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
+
+        givenThat(james).wasAbleTo(
+                clickDone,
+                acquireToken,
+                clickDone,
+                readCache);
+        int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCount.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+
+        when(james).attemptsTo(
+                clickDone,
+                acquireTokenSilent.withResourceId(mSecondaryResource),
+                clickDone,
+                readCache
+        );
+        expectedCacheCountToken  += expectedCacheCountToken/2;
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+
+    }
+
+    @Test
+    public void should_be_able_to_acquire_token_and_then_acquire_silent_frt() {
+
+        //clear cache
+        givenThat(james).wasAbleTo(clearCache, clickDone, readCache);
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(0)));
+
+        givenThat(james).wasAbleTo(
+                clickDone,
+                acquireToken
+                        .withClientId(Constants.OUTLOOK_CLIENT_ID)
+                        .withRedirectUri(Constants.OUTLOOK_REDIRECT_URI),
+                clickDone,
+                readCache);
+        int expectedCacheCountToken = james.asksFor(ExpectedCacheItemCountWithFoci.displayed());
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedCacheCountToken)));
+
+        when(james).attemptsTo(
+                clickDone,
+                acquireTokenSilent
+                        .withClientId(Constants.ONE_DRIVE_CLIENT_ID)
+                        .withRedirectUri(Constants.ONE_DRIVE_REDIRECT_URI),
+                clickDone,
+                readCache
+        );
+
+        int expectedFinalCacheCountCount = james.asksFor(ExpectedCacheItemCount.displayed()) + expectedCacheCountToken;
+        then(james).should(seeThat(TokenCacheItemCount.displayed(), is(expectedFinalCacheCountCount)));
+
+    }
+
+
+
 }
