+++ automation/src/test/java/com/microsoft/identity/common/test/automation/tasks/SignInUserADFSv2.java
@@ -30,6 +30,9 @@ import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.actions.Click;
 import net.serenitybdd.screenplay.actions.Enter;
 import net.serenitybdd.screenplay.actions.EnterValueIntoTarget;
+import net.serenitybdd.screenplay.waits.WaitUntil;
+
+import static net.serenitybdd.screenplay.matchers.WebElementStateMatchers.isVisible;
 
 public class SignInUserADFSv2 extends SignInUser {
 
@@ -37,6 +40,7 @@ public class SignInUserADFSv2 extends SignInUser {
     public <T extends Actor> void performAs(T actor) {
         User user = (User) actor;
         user.attemptsTo(
+                WaitUntil.the(SignInPage.USERNAME_FIELD, isVisible()).forNoMoreThan(10).seconds(),
                 Click.on(SignInPage.USERNAME_FIELD),
                 Enter.theValue(user.getCredential().userName).into(SignInPage.USERNAME_FIELD),
                 //Not using static method here to avoid logging the password via instrumentation... this won't show up as a step
